test : RacingGameRule 모델 추가하여 레이싱 게임 자체의 rule 에 대한 역할을 분리 및 테스트
