+++ src/main/java/racingGame/model/RacingGameRules.java
@@ -1,0 +1,19 @@
+package racingGame.model;
+
+public class RacingGameRules {
+    
+    private int moves;
+    private int cars;
+    
+    public RacingGameRules() {
+    }
+    
+    public RacingGameRules(int moves, int cars) {
+        this.moves = moves;
+        this.cars = cars;
+    }
+    
+    public RacingGameRules getRules() {
+        return null;
+    }
+}
